first draft
Questionnaire and report have been created.  The report uses no
libraries; the form uses React and jQuery (jQuery for the slider because
I refuse to use Babel and it's impossible to figure out how to use React
libraries outside of their massive and overly complicated toolchain).

It still needs to be styled.  I'm not entirely sure what it should even
look like yet.

Need to get form submission redirecting to the report page and then
it'll be ready to publish for testing.